requirements.txt
openai
python-dotenv
docx
python-docx
pandas
tqdm